backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
openai==1.3.5
requests==2.31.0
backoff==2.2.1
python-dateutil==2.8.2
django-filter==23.5
django-filter[rest_framework]
httpx>=0.25.0
